requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
pydantic
python-multipart
docling
tiktoken
python-docx
qdrant-client
google-genai
passlib
python-jose[cryptography]